refactor(transactions): extract shared ShippingField component
